requirements.txt
alembic==1.4.0
APScheduler==3.6.3
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
dialogflow==0.7.2
dominate==2.4.0
elasticsearch==7.5.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
future==0.18.2
google-api-core==1.16.0
google-auth==1.11.2
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
Mako==1.1.1
MarkupSafe==1.1.1
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
python-telegram-bot==12.4.2
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
tornado==6.0.3
tzlocal==2.0.0
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
